Pass test. Convert double to BigDecimal to better precision

+++ src/main/java/com/tmoreno/kata/coffeemachine/Order.java
@@ -1,12 +1,14 @@
 package com.tmoreno.kata.coffeemachine;
 
+import java.math.BigDecimal;
+
 public final class Order {
 
     private final Drink drink;
     private final int sugar;
-    private final double money;
+    private final BigDecimal money;
 
-    public Order(Drink drink, int sugar, double money) {
+    public Order(Drink drink, int sugar, BigDecimal money) {
         this.drink = drink;
         this.sugar = sugar;
         this.money = money;
@@ -16,7 +18,7 @@ public final class Order {
         return drink;
     }
 
-    public double getMoney() {
+    public BigDecimal getMoney() {
         return money;
     }
